Otimização no componente dos botões de cadastros + melhorias na arquitetura do projeto
