+++ lib/components/buttons/dateselect_buttons_widget.dart
@@ -1,8 +1,8 @@
 import 'package:flutter/material.dart';
 import 'package:projeto_controle_financeiro/utils/theme.dart';
 
-class DateSelectWidget extends StatelessWidget {
-  const DateSelectWidget({Key? key}) : super(key: key);
+class DateselectButtonsWidget extends StatelessWidget {
+  const DateselectButtonsWidget({Key? key}) : super(key: key);
 
   @override
   Widget build(BuildContext context) {
