+++ lib/screens/movimentacoes/page/movimentacoes_page_widget.dart
@@ -48,12 +48,14 @@ class _MovimentacoesPageWidgetState extends State<MovimentacoesPageWidget> {
       ),
       body: _widgetOptions.elementAt(_selectedIndex),
       floatingActionButton: AddButtonWidget(
-        icon1: const Icon(Icons.attach_money),
-        iconcolor1: const Color.fromARGB(255, 36, 137, 232),
-        formdialog1: FaturamentosFormWidget(),
-        icon2: const Icon(Icons.money_off),
-        iconcolor2: const Color.fromARGB(255, 255, 67, 67),
-        formdialog2: DespesasFormWidget(),
+        text1: 'Despesas',
+        icon1: const Icon(Icons.money_off),
+        iconcolor1: const Color.fromARGB(255, 255, 67, 67),
+        formdialog1: DespesasFormWidget(),
+        text2: 'Faturamentos',
+        icon2: const Icon(Icons.attach_money),
+        iconcolor2: const Color.fromARGB(255, 36, 137, 232),
+        formdialog2: FaturamentosFormWidget(),
       ),
       floatingActionButtonLocation: FloatingActionButtonLocation.centerDocked,
       bottomNavigationBar: BottomNavigationBar(
